♻️ refactor hydrogen constants and get rid of duplication
