+++ packages/hydrogen/src/types.d.ts
@@ -5,7 +5,10 @@ import type {
   FlashSessionData,
 } from '@remix-run/server-runtime';
 import type {RequestEventPayload} from './vite/request-events';
-import {CUSTOMER_ACCOUNT_SESSION_KEY, BUYER_SESSION_KEY} from './constants';
+import {
+  CUSTOMER_ACCOUNT_SESSION_KEY,
+  BUYER_SESSION_KEY,
+} from './customer/constants';
 import type {BuyerInput} from '@shopify/hydrogen-react/storefront-api-types';
 
 export interface HydrogenSessionData {
